requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
gunicorn==21.2.0
PyMySQL==1.1.0
email-validator==2.1.1
python-dotenv==1.0.1
reportlab==4.1.0
openpyxl==3.1.2
Flask-WTF==1.2.1
pandas==2.2.1  # Added pandas for excel generation
